Remove unused entries from .config-*.
I haven't actually regenerated the generated files because there's a CL
already up with an update and I don't want anyone to have to worry about
the merge conflicts. We'll get these changes "for free" when we do our
next upstream sync.

Also add a check to post_update.sh so it'll warn us of any cruft in
future.

Test: treehugger
Change-Id: If40511b2327e2dff5c3981c53d9460d474af1606